refactor account management, add new method for credential check
