+++ src/api/passport/password-local.ts
@@ -1,21 +1,21 @@
 //@see http://www.passportjs.org/packages/passport-local/
 import Local from "passport-local";
 import { debugMongo } from "~/lib/debuggers";
-import { findUserByCredentials } from "~/models/user";
-import { connectToDb } from "../middlewares/mongoConnection";
+import { findUserByCredentials } from "./account";
+import { connectToAppDb } from "../mongoose/connection";
 
+/**
+ * Passport strategy for local password based authentication
+ */
 export const localStrategy = new Local.Strategy(function (
   email,
   password,
   done
 ) {
   // TODO: logic a bit duplicated with MongoConnection handler + not tested
-  const mongoUri = process.env.MONGO_URI;
-  if (!mongoUri) throw new Error("MONGO_URI env variable is not defined");
-  const isLocalMongo = mongoUri.match(/localhost/);
-  connectToDb(mongoUri)
+  connectToAppDb()
     .then(() => {
-      debugMongo("Connected to db from route passport local strategy");
+      debugMongo("Connected to db from passport local auth strategy");
       findUserByCredentials({ email, password })
         .then((user) => {
           if (!user) {
@@ -29,13 +29,9 @@ export const localStrategy = new Local.Strategy(function (
         });
     })
     .catch((err) => {
-      console.error(
-        `\nCould not connect to Mongo database on URI ${mongoUri} during Passport local strategy.`
+      debugMongo(
+        "Could not connect to db from passport local auth strategy",
+        err
       );
-      if (isLocalMongo) {
-        console.error("Did you forget to run 'yarn run start:mongo'?\n");
-      }
-      console.error(err);
-      done(err);
     });
 });
