+++ src/api/runSeed.ts
@@ -1,19 +1,14 @@
 import mongoose from "mongoose";
 import { debugMongo } from "~/lib/debuggers";
-import { connectToDb } from "~/api/middlewares/mongoConnection";
+import { connectToAppDb } from "~/api/mongoose/connection";
 import seedDatabase from "~/api/seed";
 import { contextBase } from "~/api/context";
 
 function runSeed() {
-  const mongoUri = process.env.MONGO_URI;
-  if (!mongoUri) throw new Error("MONGO_URI env variable is not defined");
-  const isLocalMongo = mongoUri.match(/localhost/);
   // Seed in development
   // In production, we expect you to seed the database manually
   if (process.env.NODE_ENV === "development") {
-    connectToDb(mongoUri, {
-      serverSelectionTimeoutMS: isLocalMongo ? 3000 : undefined,
-    }) // fail the seed early during development
+    connectToAppDb()
       .then(() => {
         debugMongo("Connected to db, seeding admin and restaurants");
         // TODO: what is the best pattern to seed in a serverless context?
@@ -40,11 +35,8 @@ function runSeed() {
       })
       .catch((err) => {
         console.error(
-          `\nCould not connect to Mongo database on URI ${mongoUri} during seed step.`
+          `\nCould not connect to Mongo database on URI during seed step.`
         );
-        if (isLocalMongo) {
-          console.error("Did you forget to run 'yarn run start:mongo'?\n");
-        }
         console.error(err);
         process.exit(1);
       });
