+++ src/pages/api/login.ts
@@ -1,23 +1,16 @@
 import passport from "passport";
 import nextConnect from "next-connect";
-import { localStrategy } from "~/api/passport/password-local";
 import { encryptSession } from "~/api/passport/iron";
 import { setTokenCookie } from "~/api/passport/auth-cookies";
 import { NextApiRequest, NextApiResponse } from "next";
 import runSeed from "~/api/runSeed";
 
-const authenticate = (method, req, res): Promise<any> =>
-  new Promise((resolve, reject) => {
-    passport.authenticate(method, { session: false }, (error, token) => {
-      if (error) {
-        reject(error);
-      } else {
-        resolve(token);
-      }
-    })(req, res);
-  });
-
+import { localStrategy } from "~/api/passport/password-local";
+import { authenticate } from "~/api/passport/utils";
 passport.use(localStrategy);
+const authenticateWithPassword = async (req, res) => {
+  return authenticate("local", req, res);
+};
 
 // NOTE: adding NextApiRequest, NextApiResponse is required to get the right typings in next-connect
 // this is the normal behaviour
@@ -25,7 +18,7 @@ export default nextConnect<NextApiRequest, NextApiResponse>()
   .use(passport.initialize())
   .post(async (req, res) => {
     try {
-      const user = await authenticate("local", req, res);
+      const user = await authenticateWithPassword(req, res);
       // session is the payload to save in the token, it may contain basic info about the user
       const session = { ...user };
       // The token is a string with the encrypted session
